fix(save): post save hook to decrypt document
When saving an object, current implementation calls the save callback with the
encrypted mongoose document. This means that if you need to use that document
right afterword, you have to decrypt it. A better implementation is to decrypt
the document after save using a post save hook. This changes implements the
hook with tests.

No breaking changes. Now you no longer need to decrypt after save.